move some func to convert_utils
